Yargh it should work now.

+++ monkestation/code/game/objects/items/tactical_shields.dm
@@ -5,6 +5,7 @@
 /obj/item/ammo_box/advanced/s12gauge/tutel
 	name = "Tutel tactical buckler"
 	desc = "A lightweight titanium-alloy shield painted to look like wood. It has an integrated shotgun shell tube, allowing you to reload without putting down the shield."
+	icon = 'monkestation/icons/obj/weapons/shields.dmi'
 	icon_state = "buckler"
 	inhand_icon_state = "buckler"
 	custom_materials = list(/datum/material/iron= SHEET_MATERIAL_AMOUNT * 10) //Yes yes, I know, you do still repair it with titanium
@@ -12,7 +13,7 @@
 	max_ammo = 4
 	caliber = CALIBER_SHOTGUN
 	multitype = TRUE
-	block_chance = 50
+	block_chance = 75
 	max_integrity = 50
 	slot_flags = ITEM_SLOT_BACK | ITEM_SLOT_BELT
 	force = 15
@@ -39,15 +40,17 @@
 	AddComponent(/datum/component/pellet_cloud, projectile_type = /obj/projectile/bullet/shrapnel, magnitude = 1)
 	new tutel_break_leftover(get_turf(src))
 
-/obj/item/ammo_box/advanced/s12gauge/tutel/proc/on_shield_block(mob/living/carbon/human/owner, atom/movable/hitby, attack_text = "the attack", damage = 0, attack_type = MELEE_ATTACK)
-	if (atom_integrity <= damage)
-		var/turf/owner_turf = get_turf(owner)
-		owner_turf.visible_message(span_warning("[hitby] destroys [src]!"))
-		shatter(owner)
-		qdel(src)
-		return FALSE
-	take_damage(damage)
-	return TRUE
+/obj/item/ammo_box/advanced/s12gauge/tutel/hit_reaction(mob/living/carbon/human/owner, atom/movable/hitby, attack_text = "the attack", final_block_chance = 0, damage = 0, attack_type = MELEE_ATTACK)
+	. = ..()
+	if(.)
+		if (atom_integrity <= damage)
+			var/turf/owner_turf = get_turf(owner)
+			owner_turf.visible_message(span_warning("[hitby] destroys [src]!"))
+			shatter(owner)
+			qdel(src)
+			return FALSE
+		take_damage(damage)
+		return TRUE
 
 /obj/item/ammo_box/advanced/s12gauge/tutel/attackby(obj/item/attackby_item, mob/user, params)
 	if(istype(attackby_item, /obj/item/stack/sheet/mineral/titanium))
@@ -69,7 +72,7 @@
 	flags_1 = CONDUCT_1
 	force = 9
 	throwforce = 10
-	w_class = WEIGHT_CLASS_BULKY
+	w_class = WEIGHT_CLASS_NORMAL
 	custom_materials = list(/datum/material/titanium = HALF_SHEET_MATERIAL_AMOUNT + SMALL_MATERIAL_AMOUNT * 3)
 	attack_verb_continuous = list("hits", "bludgeons", "whacks", "bonks")
 	attack_verb_simple = list("hit", "bludgeon", "whack", "bonk")
@@ -89,7 +92,7 @@
 /obj/item/wirerodbroken_shield/attackby(obj/item/attacking_item, mob/user, params)
 	if(istype(attacking_item, /obj/item/stack/sheet/mineral/titanium))
 		var/datum/crafting_recipe/recipe_to_use = /datum/crafting_recipe/tutel
-		user.balloon_alert(user, "crafting spear...")
+		user.balloon_alert(user, "repairing shield...")
 		if(do_after(user, initial(recipe_to_use.time), src)) // we do initial work here to get the correct timer
 			var/obj/item/ammo_box/advanced/s12gauge/tutel/crafted_tutel = new /obj/item/ammo_box/advanced/s12gauge/tutel()
